created smurfs reducer and moved reducer
